requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
gunicorn==20.1.0 